# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /progress/{flow}:
    get:
      operationId: getFlowProgress
      tags: [progress]
      summary: Read the current visitor's saved flow progress
      description: Unlinked visitor progress remains cookie-accessible. A filmmaker visitor linked to Firebase UID requires a verified Bearer token matching that UID.
      security:
        - firebaseBearer: []
        - {}
      parameters:
        - in: path
          name: flow
          required: true
          schema:
            $ref: "#/components/schemas/Flow"
      responses:
        "200":
          description: Saved progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowProgress"
        "404":
          description: No progress saved
        "401":
          description: Verified sign-in required for an account-linked filmmaker visitor
        "403":
          description: Firebase UID does not own this visitor
        "503":
          description: Firebase verification is unavailable
  /progress:
    post:
      operationId: saveFlowProgress
      tags: [progress]
      summary: Save visitor or account-scoped in-progress answers
      description: Unlinked visitor progress remains cookie-accessible. Investor progress under a verified account is stored separately from guest progress. A filmmaker visitor linked to Firebase UID requires a matching verified Bearer token.
      security:
        - firebaseBearer: []
        - {}
      parameters:
        - name: X-MSI-Draft-Id
          in: header
          required: false
          description: Required for account-linked filmmaker visitors; must match the current filmmaker flow_progress.id.
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProgressInput"
      responses:
        "200":
          description: Saved progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowProgress"
        "400":
          description: Invalid input or missing visitor
        "401":
          description: Verified sign-in required for an account-linked filmmaker visitor
        "403":
          description: Firebase UID does not own this visitor
        "503":
          description: Firebase verification is unavailable
        "409":
          description: Draft context header is missing or does not match the current draft
...
  /filmmakers/review-checkout:
    post:
      operationId: startPitchReviewCheckout
      tags: [filmmakers]
      summary: Start pitch-specific review checkout in the explicitly configured environment
      security:
        - firebaseBearer: []
        - {}
      parameters:
        - name: X-MSI-Project-Id
          in: header
          required: false
          schema: { type: integer, minimum: 1 }
        - name: X-MSI-Checkout-Proof
          in: header
          required: false
          description: Short-lived submission proof for this pitch's review checkout when its guest visit cookie is lost.
          schema: { type: string }
      responses:
        "200":
          description: Stripe checkout URL, or already-submitted status
          content:
            application/json:
              schema:
                type: object
                required: [url, already_submitted]
                properties:
                  url: { type: ["string", "null"] }
                  already_submitted: { type: boolean }
        "400": { description: Visitor or completed pitch required }
        "401": { description: Sign-in required for an account-linked pitch }
        "409": { description: Project context changed or pitch ineligible }
        "503": { description: Checkout is gated or live Stripe verification unavailable; free pitch remains saved }
  /filmmakers/review-checkout/status:
    get:
      operationId: getPitchReviewCheckoutStatus
      tags: [filmmakers]
      summary: Reconcile payment for the current pitch and read review status
      security:
        - firebaseBearer: []
        - {}
      parameters:
        - name: X-MSI-Project-Id
          in: header
          required: false
          schema: { type: integer, minimum: 1 }
        - name: X-MSI-Checkout-Proof
          in: header
          required: false
          description: Short-lived submission proof for this pitch's review status when its guest visit cookie is lost.
          schema: { type: string }
      responses:
        "200":
          description: Review payment and moderation status
          content:
            application/json:
              schema:
                type: object
                required: [paid, pending, approved, declined]
                properties:
                  paid: { type: boolean }
                  pending: { type: boolean }
                  approved: { type: boolean }
                  declined: { type: boolean }
        "404": { description: No completed pitch found }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  schemas:
...
    Flow:
      type: string
      enum: [filmmaker, investor]
    SaveProgressInput:
      type: object
      required: [flow, last_screen, answers]
      properties:
        flow:
          $ref: "#/components/schemas/Flow"
        last_screen: { type: integer, minimum: 1, maximum: 6 }
        answers:
          type: object
          additionalProperties: true
        completed: { type: boolean }
        expected_investor_owner: { type: string, description: Investor worksheet identity captured on opening; the API rejects a save if the signed-in account changed }
...
    FlowProgress:
      type: object
      required: [flow, last_screen, answers, completed, updated_at]
      properties:
        flow:
          $ref: "#/components/schemas/Flow"
        draft_id: { type: integer, minimum: 1 }
        last_screen: { type: integer }
        answers:
          type: object
          additionalProperties: true
        completed: { type: boolean }
        updated_at: { type: string, format: date-time }
...